+++ test/integration/project-template/conftest.py
@@ -37,11 +37,6 @@ def git_path() -> str:
     return git_path
 
 
-@pytest.fixture(scope="session")
-def poetry_path() -> str:
-    result = subprocess.run(["which", "poetry"], capture_output=True, text=True)
-    poetry_path = result.stdout.strip()
-    return poetry_path
 
 
 @pytest.fixture(scope="session")
